add note about how you dont need custom checkpointer (#117)
I've noticed a couple customers have gotten confused when deploying LG
agents to LGP, because they configure custom checkpointers and stores.
But this is not supported with LGP, because these are provided out of
the box. This docs change adds a Note on step 3 of getting started which
explains this.

I don't think this will stop many cases of this confusion, so long term
maybe we should think about a simpler way to show "getting started on
LGP". This tutorial is pretty long to begin with....

---------

Co-authored-by: Kathryn May <44557882+katmayb@users.noreply.github.com>