Remove langgraph basic chatbot content from lgp docs add links (#253)
This PR removes the duplicate LangGraph Basics tutorial series from the
LGP docs. Left behind is an overview for the Build tab of the LGP site
that directly links to the LangGraph OSS version of these tutorials. It
also references that you can build with other frameworks and the
extended capabilities that LGP offers.